feat: Add CORS and Nodemailer dependencies
The code changes in the `package.json` file add the `cors` and `nodemailer` dependencies to the project. This will enable Cross-Origin Resource Sharing (CORS) and allow sending emails using Nodemailer.

This commit message follows the established convention of starting with a verb in the imperative form, followed by a concise description of the changes made.