feat: Update email service to use Mailgun for sending interest emails
The code changes in the `email.service.js` file update the email service to use Mailgun for sending interest emails. The `nodemailer` dependency is replaced with `mailgun.js` and the `sendMail` function is replaced with the `mg.messages.create` method. This change improves the reliability and deliverability of the emails.

This commit message follows the established convention of starting with a verb in the imperative form, followed by a concise description of the changes made.