refactor: Include like count in getPropertiesByOwner function
The code changes in the `property.service.js` file update the `getPropertiesByOwner` function to include the count of likes for each property. This is achieved by using the `_count` field in the `findMany` method of the `prisma.property` object. By including the like count, the response will provide more information about the popularity of each property.